# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  '/api/races/{id}':
    get:
      summary: getForId
      operationId: getForId
      description: 'Получение гонки по id'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Гонка по идентификатору получена успешно.'
                  race:
                    id: 13
                    name: 'Стрелков Денис Романович'
                    desc: 'Nihil ut aspernatur est quasi mollitia id. Expedita aut alias quo consequatur natus. Repudiandae aut eaque molestias saepe. Hic rerum hic delectus atque consequatur.'
                    is_work: true
                    date_start: '2025-02-13T04:03:54.000000Z'
                    images: null
                    contacts: null
                    position_file: null
                    results_file: null
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Гонка по идентификатору получена успешно.'
                  race:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 13
                      name:
                        type: string
                        example: 'Стрелков Денис Романович'
                      desc:
                        type: string
                        example: 'Nihil ut aspernatur est quasi mollitia id. Expedita aut alias quo consequatur natus. Repudiandae aut eaque molestias saepe. Hic rerum hic delectus atque consequatur.'
                      is_work:
                        type: boolean
                        example: true
                      date_start:
                        type: string
                        example: '2025-02-13T04:03:54.000000Z'
                      images:
                        type: string
                        example: null
                      contacts:
                        type: string
                        example: null
                      position_file:
                        type: string
                        example: null
                      results_file:
                        type: string
                        example: null
      tags:
        - Race
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: ''
                  example: 1
                  nullable: true
                appointmentUser:
                  type: boolean
                  description: ''
                  example: true
                  nullable: true
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the race.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/races/{id}/toggle-is-work':
    get:
      summary: toggleIsWork
      operationId: toggleIsWork
      description: 'Переключатель гонки, с рабочего на нерабочий и наоборот'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Состояние гонки успешно изменено на не рабочее.'
                  race:
                    id: 16
                    name: 'Дроздов Максим Владимирович'
                    desc: 'Quia qui nisi sit consequatur ex dolorem atque. Sit placeat quos dolores enim est. Voluptas ea cum ad molestiae tenetur perferendis aut.'
                    is_work: false
                    date_start: '2025-02-13T09:45:51.000000Z'
                    images: null
                    contacts: null
                    position_file: null
                    results_file: null
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Состояние гонки успешно изменено на не рабочее.'
                  race:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 16
                      name:
                        type: string
                        example: 'Дроздов Максим Владимирович'
                      desc:
                        type: string
                        example: 'Quia qui nisi sit consequatur ex dolorem atque. Sit placeat quos dolores enim est. Voluptas ea cum ad molestiae tenetur perferendis aut.'
                      is_work:
                        type: boolean
                        example: false
                      date_start:
                        type: string
                        example: '2025-02-13T09:45:51.000000Z'
                      images:
                        type: string
                        example: null
                      contacts:
                        type: string
                        example: null
                      position_file:
                        type: string
                        example: null
                      results_file:
                        type: string
                        example: null
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Доступ запрещен.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Доступ запрещен.'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Запрашиваемый ресурс не найден.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Запрашиваемый ресурс не найден.'
      tags:
        - Race
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the race.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/tracks/{id}':
    get:
      summary: getForId
      operationId: getForId
      description: 'Получение трассы по id'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Трек по идентификатору получен успешно.'
                  track:
                    id: 13
                    name: Электромонтёр
                    address: '356516, Курганская область, город Шатура, ул. Сталина, 24'
                    images: null
                    latitude: 0
                    longitude: 0
                    level:
                      name: Средняя
                      color: '#EF7C08'
                    desc: 'Totam voluptates voluptatem veniam laborum. Laboriosam hic ea necessitatibus veritatis illo ea. Quaerat eos nulla rerum cumque reprehenderit sit ad. Earum quidem qui laboriosam magnam quia vel ipsum aut. Cupiditate quo sint nemo et maiores rerum. Voluptatem id ex est eveniet qui. Ab itaque iusto alias quas ut voluptas quia. Sapiente et alias laboriosam. Rerum minima aperiam at quibusdam. Eum ex doloremque et quae quo numquam nostrum. Neque cumque sint sed.'
                    length: 974
                    turns: 12
                    free: false
                    is_work: false
                    spec: null
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Трек по идентификатору получен успешно.'
                  track:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 13
                      name:
                        type: string
                        example: Электромонтёр
                      address:
                        type: string
                        example: '356516, Курганская область, город Шатура, ул. Сталина, 24'
                      images:
                        type: string
                        example: null
                      latitude:
                        type: integer
                        example: 0
                      longitude:
                        type: integer
                        example: 0
                      level:
                        type: object
                        properties:
                          name:
                            type: string
                            example: Средняя
                          color:
                            type: string
                            example: '#EF7C08'
                      desc:
                        type: string
                        example: 'Totam voluptates voluptatem veniam laborum. Laboriosam hic ea necessitatibus veritatis illo ea. Quaerat eos nulla rerum cumque reprehenderit sit ad. Earum quidem qui laboriosam magnam quia vel ipsum aut. Cupiditate quo sint nemo et maiores rerum. Voluptatem id ex est eveniet qui. Ab itaque iusto alias quas ut voluptas quia. Sapiente et alias laboriosam. Rerum minima aperiam at quibusdam. Eum ex doloremque et quae quo numquam nostrum. Neque cumque sint sed.'
                      length:
                        type: integer
                        example: 974
                      turns:
                        type: integer
                        example: 12
                      free:
                        type: boolean
                        example: false
                      is_work:
                        type: boolean
                        example: false
                      spec:
                        type: string
                        example: null
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Запрашиваемый ресурс не найден.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Запрашиваемый ресурс не найден.'
      tags:
        - Track
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the track.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/users/{id}':
    get:
      summary: getForId
      operationId: getForId
      description: 'Получить пользователя по id'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Пользователь получен успешно.'
                  user:
                    id: 1107
                    name: 'Абрамова Анфиса Владимировна'
                    avatar: null
                    personal: []
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Пользователь получен успешно.'
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1107
                      name:
                        type: string
                        example: 'Абрамова Анфиса Владимировна'
                      avatar:
                        type: string
                        example: null
                      personal:
                        type: array
                        example: []
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Запрашиваемый ресурс не найден.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Запрашиваемый ресурс не найден.'
      tags:
        - User
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...